Comments: Split tests, added extra archive/reference tests
